requirements.txt
# Google Sheets API
gspread>=6.0.0
google-auth>=2.0.0
# Core dependencies for algorithmic trading system
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.25.0
scikit-learn>=1.3.0
requests>=2.28.0